Fix: Circular defcustoms
Fixes #1 (probably, maybe, who knows).

This is the only way I have been able to find that seems to fix this
problem.  I tried using ":set-after '(magit-todos-ignored-keywords)"
on `magit-todos-keywords', but it had no effect.  I looked in the
manual, which seems to suggest that using ":initialize
'custom-initialize-safe-set" might fix it--but
`custom-initialize-safe-set' is no longer to be found in the Emacs
source tree.  It was committed in 2005, and now it's gone, but the
manual still mentions it. ???

Anyway, it doesn't seem like this should be that unusual of a thing to
do, having one defcustom var that interacts with the value of another
one when it's set, but since the two ways that the manual seems to
suggest to handle it don't work or don't exist anymore...