Add: Peek command
See #13, #15.

Also turn --goto-item into -jump-to-item.

I learned (or did I?) something very confusing about Emacs: it is
apparently not possible to move point in a buffer from within
with-current-buffer when that buffer is already displayed in another
window.  You must switch to that buffer.  Or something like that,
anyway.  The way that seemed most natural to do this, having a
function that would return a buffer with its point on the item, didn't
work when the buffer was already visible in another window: the point
simply didn't move.  Maybe there's a better way, but at least this
works.