Fix: Let async-when-done kill scanner process buffers
This seems to be the correct thing to do, and it seems to work
correctly.  But since I haven't had any reports of errors from other
users, and I'm not completely sure this won't cause any problems, I'm
putting this in the next minor version to allow for extra testing.