Fix: async deleted buffer/process bugs
Copying async-when-done into this project with minor changes isn't an
ideal solution, but it seems to work around the problem.  Been testing
it for several weeks without issue, so merging now.  If it's a bug in
async, maybe it can be fixed upstream in the future.  If not, maybe a
better solution than this will be found.

Fixes #58, I think.  Please let me know if it doesn't.