scanner: first version of more generic matcher composition
Instead of hardcoding two different versions of the search and the
result regexp, we derive both from four different patterns that are
simply concatenated:

   keyword-prefix + keyword + keyword-suffix + description-suffix

For the result regexp, we transform the regexp parse tree to make the
most top-level groups in each component a named group. For example, if
the keyword-prefix is

   (optional (group (1+ "*")) blank)

the tree transformer returns an regular expression that looks like:

   (optional (group-n 1 (1+ "*")) blank)

With this change, the scanner code become more clear and flexible in
terms of user customizability. Furthermore, I could remove two HACKs
for the suffix code.