Fix error caused by display-warning returning t
Fix "error in process sentinel: Wrong type argument: magit-todos-item,
t" occurring when there's a stack overflow in the regexp matcher,
which is caught and then a warning displayed, but since
display-warning returns a truthy value (t) this truthy value is
treated as a valid todo item by the enclosing --when-let.  The error
is then thrown later when Emacs tries to write its the TODO item in
the magit status buffer.