Do not set parent keymap of magit-todos-section-map
This is not necessary to make the regular magit-status keys work in
the TODOs section.

In fact, it causes a bug when used with evil-collection, which uses x
instead of k for delete (so that k can move the cursor up one line).
Setting the parent keymap unnecessarily confuses evil-collection and
causes k to attempt to delete the "thing" at point instead of
scrolling up.